Estimate a Robust Bayesian T-Test

RoBTT is used to estimate a Robust Bayesian T-Test. The input either requires the vector of observations for each group, x1, x2, or the summary statistics (in case only the "normal" likelihood is used).

RoBTT(
  x1 = NULL,
  x2 = NULL,
  mean1 = NULL,
  mean2 = NULL,
  sd1 = NULL,
  sd2 = NULL,
  N1 = NULL,
  N2 = NULL,
  prior_delta = prior(distribution = "cauchy", parameters = list(location = 0, scale =
    sqrt(2)/2)),
  prior_rho = prior(distribution = "beta", parameters = list(alpha = 1, beta = 1)),
  prior_nu = prior(distribution = "exp", parameters = list(rate = 1)),
  prior_delta_null = prior(distribution = "spike", parameters = list(location = 0)),
  prior_rho_null = prior(distribution = "spike", parameters = list(location = 0.5)),
  prior_nu_null = NULL,
  likelihood = c("normal", if (!is.null(prior_nu)) "t"),
  chains = 4,
  iter = 10000,
  warmup = 5000,
  thin = 1,
  parallel = FALSE,
  control = set_control(),
  convergence_checks = set_convergence_checks(),
  save = "all",
  seed = NULL,
  silent = TRUE,
  ...
)

Arguments

x1

vector of observations of the first group

x2

vector of observations of the second group

mean1

mean of the first group

mean2

mean of the first group

sd1

standard deviation of the first group

sd2

standard deviation of the first group

N1

sample size of the first group

N2

sample size of the first group

prior_delta

prior distributions for the effect size delta parameter that will be treated as belonging to the alternative hypothesis. Defaults to prior(distribution = "Cauchy", parameters = list(location = 0, scale = sqrt(2)/2)).

prior_rho

prior distributions for the precision allocation rho parameter that will be treated as belonging to the alternative hypothesis. Defaults to prior(distribution = "beta", parameters = list(alpha = 1, beta = 1)).

prior_nu

prior distribution for the degrees of freedom + 2 nu parameter that will be treated as belonging to the alternative hypothesis. Defaults to prior(distribution = "exp", parameters = list(rate = 1)).

prior_delta_null

prior distribution for the delta parameter that will be treated as belonging to the null hypothesis. Defaults to point distribution with location at 0 ( prior(distribution = "point", parameters = list(location = 0))).

prior_rho_null

prior distribution for the rho parameter that will be treated as belonging to the null hypothesis. Defaults to point distribution with location at 0.5 ( prior(distribution = "point", parameters = list(location = 0.5))).

prior_nu_null

prior distribution for the nu parameter that will be treated as belonging to the null hypothesis. Defaults to NULL ( NULL, i.e., normal distribution).

likelihood

types of likelihood to be used. Defaults to c("normal", "t") for a normal and t likelihoods.

chains

a number of chains of the MCMC algorithm.

iter

a number of sampling iterations of the MCMC algorithm. Defaults to 10000, with a minimum of 4000.

warmup

a number of warmup iterations of the MCMC algorithm. Defaults to 5000.

thin

a thinning of the chains of the MCMC algorithm. Defaults to 1.

parallel

whether the individual models should be fitted in parallel. Defaults to FALSE. The implementation is not completely stable and might cause a connection error.

control

allows to pass control settings with the set_control() function. See ?set_control for options and default settings.

convergence_checks

automatic convergence checks to assess the fitted models, passed with set_convergence_checks() function. See ?set_convergence_checks for options and default settings.

save

whether all models posterior distributions should be kept after obtaining a model-averaged result. Defaults to "all" which does not remove anything. Set to "min" to significantly reduce the size of final object, however, some model diagnostics and further manipulation with the object will not be possible.

seed

a seed to be set before model fitting, marginal likelihood computation, and posterior mixing for reproducibility of results. Defaults to NULL - no seed is set.

silent

whether all print messages regarding the fitting process should be suppressed. Defaults to TRUE. Note that parallel = TRUE also suppresses all messages.

...

additional arguments.

Value

RoBTT returns an object of class

"RoBTT".

Details

See Maier et al. (2022) for more details regarding the methodology.

Generic summary.RoBTT(), print.RoBTT(), and plot.RoBTT() functions are provided to facilitate manipulation with the ensemble.

References

Maier M, Bartoš F, Quintana DS, van den Bergh D, Marsman M, Ly A, Wagenmakers E (2022). “Model-averaged Bayesian t-tests.” doi:10.31234/osf.io/d5zwc , PsyArxiv Preprint.

See also

Examples

# using the example data from Darwin
data("fertilization", package = "RoBTT")
fit <- RoBTT(
  x1       = fertilization$Self,
  x2       = fertilization$Crossed,
  prior_delta = prior("cauchy", list(0, 1/sqrt(2))),
  prior_rho   = prior("beta",   list(3, 3)),
  likelihood  = "normal",
  seed        = 1, 
  chains      = 1,
  warmup      = 1000,
  iter        = 2000,
  control     = set_control(adapt_delta = 0.95)
)
#> Chain 1: 
#> Chain 1: Gradient evaluation took 1.2e-05 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 0.12 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: 
#> Chain 1:  Elapsed Time: 0.022128 seconds (Warm-up)
#> Chain 1:                0.017784 seconds (Sampling)
#> Chain 1:                0.039912 seconds (Total)
#> Chain 1: 
#> 0 of 1000 iterations ended with a divergence.
#> Chain 1: 
#> Chain 1: Gradient evaluation took 1.7e-05 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 0.17 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: 
#> Chain 1:  Elapsed Time: 0.034141 seconds (Warm-up)
#> Chain 1:                0.041978 seconds (Sampling)
#> Chain 1:                0.076119 seconds (Total)
#> Chain 1: 
#> 0 of 1000 iterations ended with a divergence.
#> Chain 1: 
#> Chain 1: Gradient evaluation took 1.5e-05 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 0.15 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: 
#> Chain 1:  Elapsed Time: 0.027182 seconds (Warm-up)
#> Chain 1:                0.025181 seconds (Sampling)
#> Chain 1:                0.052363 seconds (Total)
#> Chain 1: 
#> 0 of 1000 iterations ended with a divergence.
#> Chain 1: 
#> Chain 1: Gradient evaluation took 2e-05 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 0.2 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: 
#> Chain 1:  Elapsed Time: 0.044123 seconds (Warm-up)
#> Chain 1:                0.041091 seconds (Sampling)
#> Chain 1:                0.085214 seconds (Total)
#> Chain 1: 
#> 0 of 1000 iterations ended with a divergence.
#> Warning: Model (2): Minimum effective sample size was low (325).
#> Warning: Model (4): Minimum effective sample size was low (469).

# summary can provide many details about the model
summary(fit)
#> Call:
#> RoBTT(x1 = fertilization$Self, x2 = fertilization$Crossed, prior_delta = prior("cauchy", 
#>     list(0, 1/sqrt(2))), prior_rho = prior("beta", list(3, 3)), 
#>     likelihood = "normal", chains = 1, iter = 2000, warmup = 1000, 
#>     control = set_control(adapt_delta = 0.95), seed = 1)
#> 
#> Robust Bayesian t-test
#> Components summary:
#>               Models Prior prob. Post. prob. Inclusion BF
#> Effect           2/4       0.500       0.750        3.002
#> Heterogeneity    2/4       0.500       0.709        2.431
#> 
#> Model-averaged estimates:
#>        Mean Median 0.025 0.975
#> delta 0.540  0.546 0.000 1.425
#> rho   0.632  0.637 0.469 0.849
#> Model (2): Minimum effective sample size was low (325).
#> Model (4): Minimum effective sample size was low (469).