Updates a fitted RoBTT object

update.RoBTT can be used to

change the prior odds of fitted models by specifying a vector prior_weights of the same length as the fitted models,

refitting models that failed to converge with updated settings of control parameters,

or changing the convergence criteria and recalculating the ensemble results by specifying new control argument and setting refit_failed == FALSE.

# S3 method for RoBTT
update(
  object,
  refit_failed = TRUE,
  prior_weights = NULL,
  chains = NULL,
  iter = NULL,
  warmup = NULL,
  thin = NULL,
  parallel = NULL,
  control = NULL,
  convergence_checks = NULL,
  save = "all",
  seed = NULL,
  silent = TRUE,
  ...
)

Arguments

object

a fitted RoBTT object

refit_failed

whether failed models should be refitted. Relevant only prior_weights are not supplied. Defaults to TRUE.

prior_weights

either a single value specifying prior model weight of a newly specified model using priors argument, or a vector of the same length as already fitted models to update their prior weights.

chains

a number of chains of the MCMC algorithm.

iter

a number of sampling iterations of the MCMC algorithm. Defaults to 10000, with a minimum of 4000.

warmup

a number of warmup iterations of the MCMC algorithm. Defaults to 5000.

thin

a thinning of the chains of the MCMC algorithm. Defaults to 1.

parallel

whether the individual models should be fitted in parallel. Defaults to FALSE. The implementation is not completely stable and might cause a connection error.

control

allows to pass control settings with the set_control() function. See ?set_control for options and default settings.

convergence_checks

automatic convergence checks to assess the fitted models, passed with set_convergence_checks() function. See ?set_convergence_checks for options and default settings.

save

whether all models posterior distributions should be kept after obtaining a model-averaged result. Defaults to "all" which does not remove anything. Set to "min" to significantly reduce the size of final object, however, some model diagnostics and further manipulation with the object will not be possible.

seed

a seed to be set before model fitting, marginal likelihood computation, and posterior mixing for reproducibility of results. Defaults to NULL - no seed is set.

silent

whether all print messages regarding the fitting process should be suppressed. Defaults to TRUE. Note that parallel = TRUE also suppresses all messages.

...

additional arguments.

Value

RoBTT returns an object of class 'RoBTT'.

Details

See RoBTT() for more details.

See also