Summarize fitted 'RoBTT' object

summary.RoBTT creates summary tables for a RoBTT object.

# S3 method for RoBTT
summary(
  object,
  type = "ensemble",
  conditional = FALSE,
  group_estimates = FALSE,
  probs = c(0.025, 0.975),
  logBF = FALSE,
  BF01 = FALSE,
  short_name = FALSE,
  remove_spike_0 = FALSE,
  ...
)

Arguments

object

a fitted 'RoBTT' object

type

whether to show the overall 'RoBTT' results ("ensemble"), an overview of the individual models ("models"), an overview of the individual models MCMC diagnostics ("diagnostics"), or a detailed summary of the individual models ("individual"). Can be abbreviated to first letters.

conditional

show the conditional estimates (assuming that the alternative is true). Defaults to FALSE. Only available for type == "conditional".

group_estimates

show the model-averaged mean and standard deviation estimates for each group.

probs

quantiles of the posterior samples to be displayed. Defaults to c(.025, .975)

logBF

show log of Bayes factors. Defaults to FALSE.

BF01

show Bayes factors in support of the null hypotheses. Defaults to FALSE.

short_name

whether priors names should be shortened to the first (couple) of letters. Defaults to FALSE.

remove_spike_0

whether spike prior distributions with location at zero should be omitted from the summary. Defaults to FALSE.

...

additional arguments

Value

summary.RoBTT returns a list of tables of class 'BayesTools_table'.

See also

Examples

# using the example data from Darwin
data("fertilization", package = "RoBTT")
fit <- RoBTT(
  x1       = fertilization$Self,
  x2       = fertilization$Crossed,
  prior_delta = prior("cauchy", list(0, 1/sqrt(2))),
  prior_rho   = prior("beta",   list(3, 3)),
  seed        = 1, 
  chains      = 1,
  warmup      = 1000,
  iter        = 2000,
  control     = set_control(adapt_delta = 0.95)
)
#> Chain 1: 
#> Chain 1: Gradient evaluation took 1.1e-05 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 0.11 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: 
#> Chain 1:  Elapsed Time: 0.029 seconds (Warm-up)
#> Chain 1:                0.025 seconds (Sampling)
#> Chain 1:                0.054 seconds (Total)
#> Chain 1: 
#> 0 of 1000 iterations ended with a divergence.
#> Chain 1: 
#> Chain 1: Gradient evaluation took 1.9e-05 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 0.19 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: 
#> Chain 1:  Elapsed Time: 0.071 seconds (Warm-up)
#> Chain 1:                0.066 seconds (Sampling)
#> Chain 1:                0.137 seconds (Total)
#> Chain 1: 
#> 0 of 1000 iterations ended with a divergence.
#> Chain 1: 
#> Chain 1: Gradient evaluation took 1.8e-05 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 0.18 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: 
#> Chain 1:  Elapsed Time: 0.048 seconds (Warm-up)
#> Chain 1:                0.053 seconds (Sampling)
#> Chain 1:                0.101 seconds (Total)
#> Chain 1: 
#> 0 of 1000 iterations ended with a divergence.
#> Chain 1: 
#> Chain 1: Gradient evaluation took 2.3e-05 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 0.23 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: 
#> Chain 1:  Elapsed Time: 0.106 seconds (Warm-up)
#> Chain 1:                0.109 seconds (Sampling)
#> Chain 1:                0.215 seconds (Total)
#> Chain 1: 
#> 0 of 1000 iterations ended with a divergence.
#> Chain 1: 
#> Chain 1: Gradient evaluation took 1.6e-05 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 0.16 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: 
#> Chain 1:  Elapsed Time: 0.035 seconds (Warm-up)
#> Chain 1:                0.033 seconds (Sampling)
#> Chain 1:                0.068 seconds (Total)
#> Chain 1: 
#> 0 of 1000 iterations ended with a divergence.
#> Chain 1: 
#> Chain 1: Gradient evaluation took 1.9e-05 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 0.19 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: 
#> Chain 1:  Elapsed Time: 0.111 seconds (Warm-up)
#> Chain 1:                0.116 seconds (Sampling)
#> Chain 1:                0.227 seconds (Total)
#> Chain 1: 
#> 0 of 1000 iterations ended with a divergence.
#> Chain 1: 
#> Chain 1: Gradient evaluation took 1.7e-05 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 0.17 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: 
#> Chain 1:  Elapsed Time: 0.055 seconds (Warm-up)
#> Chain 1:                0.054 seconds (Sampling)
#> Chain 1:                0.109 seconds (Total)
#> Chain 1: 
#> 0 of 1000 iterations ended with a divergence.
#> Chain 1: 
#> Chain 1: Gradient evaluation took 2.3e-05 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 0.23 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: 
#> Chain 1:  Elapsed Time: 0.177 seconds (Warm-up)
#> Chain 1:                0.139 seconds (Sampling)
#> Chain 1:                0.316 seconds (Total)
#> Chain 1: 
#> 0 of 1000 iterations ended with a divergence.
#> Warning: Model (2): Minimum effective sample size was low (95).
#> Warning: Model (3): Minimum effective sample size was low (325).
#> Warning: Model (4): Minimum effective sample size was low (160).
#> Warning: Model (6): Minimum effective sample size was low (220).
#> Warning: Model (7): Minimum effective sample size was low (469).
#> Warning: There were another 0 warnings. To see all warnings call 'check_RoBTT(fit)'.

# summary can provide many details about the model
summary(fit)
#> Call:
#> RoBTT(x1 = fertilization$Self, x2 = fertilization$Crossed, prior_delta = prior("cauchy", 
#>     list(0, 1/sqrt(2))), prior_rho = prior("beta", list(3, 3)), 
#>     chains = 1, iter = 2000, warmup = 1000, control = set_control(adapt_delta = 0.95), 
#>     seed = 1)
#> 
#> Robust Bayesian t-test
#> Components summary:
#>               Models Prior prob. Post. prob. Inclusion BF
#> Effect           4/8       0.500       0.975       39.159
#> Heterogeneity    4/8       0.500       0.451        0.822
#> Outliers         4/8       0.500       0.923       11.912
#> 
#> Model-averaged estimates:
#>        Mean Median 0.025 0.975
#> delta 0.870  0.854 0.000 1.695
#> rho   0.532  0.500 0.335 0.787
#> nu      Inf  2.602 2.047   Inf
#> Model (2): Minimum effective sample size was low (95).
#> Model (3): Minimum effective sample size was low (325).
#> Model (4): Minimum effective sample size was low (160).
#> Model (6): Minimum effective sample size was low (220).
#> Model (7): Minimum effective sample size was low (469).
#> There were another 0 warnings. To see all warnings call 'check_RoBTT(fit)'.

# estimates from the conditional models can be obtained with
summary(fit, conditional = TRUE)
#> Call:
#> RoBTT(x1 = fertilization$Self, x2 = fertilization$Crossed, prior_delta = prior("cauchy", 
#>     list(0, 1/sqrt(2))), prior_rho = prior("beta", list(3, 3)), 
#>     chains = 1, iter = 2000, warmup = 1000, control = set_control(adapt_delta = 0.95), 
#>     seed = 1)
#> 
#> Robust Bayesian t-test
#> Components summary:
#>               Models Prior prob. Post. prob. Inclusion BF
#> Effect           4/8       0.500       0.975       39.159
#> Heterogeneity    4/8       0.500       0.451        0.822
#> Outliers         4/8       0.500       0.923       11.912
#> 
#> Model-averaged estimates:
#>        Mean Median 0.025 0.975
#> delta 0.870  0.854 0.000 1.695
#> rho   0.532  0.500 0.335 0.787
#> nu      Inf  2.602 2.047   Inf
#> Model (2): Minimum effective sample size was low (95).
#> Model (3): Minimum effective sample size was low (325).
#> Model (4): Minimum effective sample size was low (160).
#> Model (6): Minimum effective sample size was low (220).
#> Model (7): Minimum effective sample size was low (469).
#> There were another 0 warnings. To see all warnings call 'check_RoBTT(fit)'.
#> 
#> Conditional estimates:
#>        Mean Median 0.025 0.975
#> delta 0.890  0.862 0.254 1.697
#> rho   0.571  0.575 0.294 0.825
#> nu    2.792  2.538 2.044 5.032
#> Model (2): Minimum effective sample size was low (95).
#> Model (3): Minimum effective sample size was low (325).
#> Model (4): Minimum effective sample size was low (160).
#> Model (6): Minimum effective sample size was low (220).
#> Model (7): Minimum effective sample size was low (469).
#> There were another 0 warnings. To see all warnings call 'check_RoBTT(fit)'.

# overview of the models and their prior and posterior probability, marginal likelihood,
# and inclusion Bayes factor can be obtained with
summary(fit, type = "models")
#> Call:
#> RoBTT(x1 = fertilization$Self, x2 = fertilization$Crossed, prior_delta = prior("cauchy", 
#>     list(0, 1/sqrt(2))), prior_rho = prior("beta", list(3, 3)), 
#>     chains = 1, iter = 2000, warmup = 1000, control = set_control(adapt_delta = 0.95), 
#>     seed = 1)
#> 
#> Robust Bayesian t-test
#> Models overview:
#>  Model Distribution   Prior delta    Prior rho    Prior nu    Prior prob.
#>      1       normal        Spike(0) Spike(0.5)                      0.125
#>      2            t        Spike(0) Spike(0.5) Exponential(1)       0.125
#>      3       normal        Spike(0) Beta(3, 3)                      0.125
#>      4            t        Spike(0) Beta(3, 3) Exponential(1)       0.125
#>      5       normal Cauchy(0, 0.71) Spike(0.5)                      0.125
#>      6            t Cauchy(0, 0.71) Spike(0.5) Exponential(1)       0.125
#>      7       normal Cauchy(0, 0.71) Beta(3, 3)                      0.125
#>      8            t Cauchy(0, 0.71) Beta(3, 3) Exponential(1)       0.125
#>  log(marglik) Post. prob. Inclusion BF
#>        -76.82       0.006        0.040
#>        -77.79       0.002        0.015
#>        -75.96       0.014        0.097
#>        -77.36       0.003        0.024
#>        -75.75       0.017        0.120
#>        -72.31       0.524        7.710
#>        -74.85       0.041        0.301
#>        -72.60       0.393        4.530
#> Model (2): Minimum effective sample size was low (95).
#> Model (3): Minimum effective sample size was low (325).
#> Model (4): Minimum effective sample size was low (160).
#> Model (6): Minimum effective sample size was low (220).
#> Model (7): Minimum effective sample size was low (469).
#> There were another 0 warnings. To see all warnings call 'check_RoBTT(fit)'.

# diagnostics overview, containing the maximum R-hat, minimum ESS, maximum MCMC error, and
# maximum MCMC error / sd across parameters for each individual model can be obtained with
summary(fit, type = "diagnostics")
#> Call:
#> RoBTT(x1 = fertilization$Self, x2 = fertilization$Crossed, prior_delta = prior("cauchy", 
#>     list(0, 1/sqrt(2))), prior_rho = prior("beta", list(3, 3)), 
#>     chains = 1, iter = 2000, warmup = 1000, control = set_control(adapt_delta = 0.95), 
#>     seed = 1)
#> 
#> Robust Bayesian t-test
#> Diagnostics overview:
#>  Model Distribution   Prior delta    Prior rho    Prior nu    max[error(MCMC)]
#>      1       normal        Spike(0) Spike(0.5)                         0.13670
#>      2            t        Spike(0) Spike(0.5) Exponential(1)         75.26249
#>      3       normal        Spike(0) Beta(3, 3)                         0.14611
#>      4            t        Spike(0) Beta(3, 3) Exponential(1)          4.77964
#>      5       normal Cauchy(0, 0.71) Spike(0.5)                         0.09697
#>      6            t Cauchy(0, 0.71) Spike(0.5) Exponential(1)          2.52735
#>      7       normal Cauchy(0, 0.71) Beta(3, 3)                         0.10366
#>      8            t Cauchy(0, 0.71) Beta(3, 3) Exponential(1)          2.45693
#>  max[error(MCMC)/SD] min(ESS) max(R-hat)
#>                0.043      549      1.000
#>                0.103       95      1.007
#>                0.055      325      1.007
#>                0.079      160      1.001
#>                0.036      787      1.001
#>                0.067      220      1.006
#>                0.046      469      1.000
#>                0.078      166      1.009
#> Model (2): Minimum effective sample size was low (95).
#> Model (3): Minimum effective sample size was low (325).
#> Model (4): Minimum effective sample size was low (160).
#> Model (6): Minimum effective sample size was low (220).
#> Model (7): Minimum effective sample size was low (469).
#> There were another 0 warnings. To see all warnings call 'check_RoBTT(fit)'.

# summary of individual models and their parameters can be further obtained by
summary(fit, type = "individual")
#> Call:
#> RoBTT(x1 = fertilization$Self, x2 = fertilization$Crossed, prior_delta = prior("cauchy", 
#>     list(0, 1/sqrt(2))), prior_rho = prior("beta", list(3, 3)), 
#>     chains = 1, iter = 2000, warmup = 1000, control = set_control(adapt_delta = 0.95), 
#>     seed = 1)
#> 
#> Robust Bayesian t-test                                                                
#>  Model               1             Parameter prior distributions
#>  Prior prob.     0.125                        delta ~ Spike(0)  
#>  log(marglik)   -76.82                          rho ~ Spike(0.5)
#>  Post. prob.     0.006                                          
#>  Inclusion BF    0.040                                          
#> 
#> Parameter estimates:
#>                Mean    SD    lCI Median    uCI error(MCMC) error(MCMC)/SD ESS
#> mu           18.890 0.590 17.746 18.876 20.058     0.02222          0.038 706
#> sigma2       10.981 3.203  6.380 10.348 18.176     0.13670          0.043 549
#> pooled_sigma  3.281 0.463  2.526  3.217  4.263     0.01960          0.042 558
#> sigma_i[1]    3.281 0.463  2.526  3.217  4.263     0.01960          0.042 558
#> sigma_i[2]    3.281 0.463  2.526  3.217  4.263     0.01960          0.042 558
#> mu_i[1]      18.890 0.590 17.746 18.876 20.058     0.02222          0.038 706
#> mu_i[2]      18.890 0.590 17.746 18.876 20.058     0.02222          0.038 706
#>              R-hat
#> mu           1.000
#> sigma2       1.000
#> pooled_sigma 1.000
#> sigma_i[1]   1.000
#> sigma_i[2]   1.000
#> mu_i[1]      1.000
#> mu_i[2]      1.000
#> 
#>                                                                 
#>  Model               2             Parameter prior distributions
#>  Prior prob.     0.125                    delta ~ Spike(0)      
#>  log(marglik)   -77.79                      rho ~ Spike(0.5)    
#>  Post. prob.     0.002                       nu ~ Exponential(1)
#>  Inclusion BF    0.015                                          
#> 
#> Parameter estimates:
#>                 Mean       SD    lCI Median      uCI error(MCMC) error(MCMC)/SD
#> mu            19.242    0.580 18.086 19.249   20.368     0.02700          0.047
#> sigma2       165.426 1011.385  8.400 19.935 1277.368    75.26249          0.074
#> nu_p[1]        1.325    1.160  0.014  1.084    4.075     0.07076          0.061
#> pooled_sigma   7.220   10.650  2.898  4.465   35.740     1.09370          0.103
#> sigma_i[1]     7.220   10.650  2.898  4.465   35.740     1.09370          0.103
#> sigma_i[2]     7.220   10.650  2.898  4.465   35.740     1.09370          0.103
#> scale_i[1]     2.593    0.497  1.737  2.548    3.777     0.01455          0.029
#> scale_i[2]     2.593    0.497  1.737  2.548    3.777     0.01455          0.029
#> mu_i[1]       19.242    0.580 18.086 19.249   20.368     0.02700          0.047
#> mu_i[2]       19.242    0.580 18.086 19.249   20.368     0.02700          0.047
#> nu             3.325    1.160  2.014  3.084    6.075     0.07076          0.061
#>               ESS R-hat
#> mu            461 1.003
#> sigma2        181 0.999
#> nu_p[1]       269 1.003
#> pooled_sigma   95 1.007
#> sigma_i[1]     95 1.007
#> sigma_i[2]     95 1.007
#> scale_i[1]   1168 0.999
#> scale_i[2]   1168 0.999
#> mu_i[1]       461 1.003
#> mu_i[2]       461 1.003
#> nu            269 1.003
#> Minimum effective sample size was low (95).
#> 
#>                                                                 
#>  Model               3             Parameter prior distributions
#>  Prior prob.     0.125                        delta ~ Spike(0)  
#>  log(marglik)   -75.96                          rho ~ Beta(3, 3)
#>  Post. prob.     0.014                                          
#>  Inclusion BF    0.097                                          
#> 
#> Parameter estimates:
#>                Mean    SD    lCI Median    uCI error(MCMC) error(MCMC)/SD ESS
#> mu           18.348 0.606 17.248 18.331 19.571     0.02630          0.043 532
#> sigma2        8.556 2.835  4.739  8.099 15.212     0.14611          0.052 376
#> rho[1]        0.698 0.109  0.436  0.711  0.872     0.00598          0.055 335
#> pooled_sigma  3.273 0.483  2.526  3.195  4.392     0.01858          0.038 676
#> sigma_i[1]    2.489 0.552  1.689  2.404  3.754     0.03060          0.055 325
#> sigma_i[2]    3.860 0.698  2.784  3.752  5.492     0.02363          0.034 873
#> mu_i[1]      18.348 0.606 17.248 18.331 19.571     0.02630          0.043 532
#> mu_i[2]      18.348 0.606 17.248 18.331 19.571     0.02630          0.043 532
#>              R-hat
#> mu           1.001
#> sigma2       1.007
#> rho[1]       1.003
#> pooled_sigma 0.999
#> sigma_i[1]   1.006
#> sigma_i[2]   0.999
#> mu_i[1]      1.001
#> mu_i[2]      1.001
#> Minimum effective sample size was low (325).
#> 
#>                                                                 
#>  Model               4             Parameter prior distributions
#>  Prior prob.     0.125                    delta ~ Spike(0)      
#>  log(marglik)   -77.36                      rho ~ Beta(3, 3)    
#>  Post. prob.     0.003                       nu ~ Exponential(1)
#>  Inclusion BF    0.024                                          
#> 
#> Parameter estimates:
#>                Mean     SD    lCI Median     uCI error(MCMC) error(MCMC)/SD ESS
#> mu           18.791  0.654 17.533 18.769  20.107     0.03100          0.047 446
#> sigma2       30.483 70.572  5.633 15.410 138.721     4.77964          0.068 218
#> rho[1]        0.666  0.143  0.349  0.680   0.893     0.00727          0.051 388
#> nu_p[1]       1.532  1.223  0.078  1.225   4.674     0.06727          0.055 331
#> pooled_sigma  5.292  3.177  2.877  4.411  13.469     0.25124          0.079 160
#> sigma_i[1]    4.222  2.700  1.874  3.455  10.914     0.21074          0.078 164
#> sigma_i[2]    6.062  3.786  3.067  5.099  15.020     0.29250          0.077 168
#> scale_i[1]    2.145  0.589  1.209  2.076   3.478     0.02667          0.045 488
#> scale_i[2]    3.120  0.757  1.872  3.039   4.780     0.02940          0.039 662
#> mu_i[1]      18.791  0.654 17.533 18.769  20.107     0.03100          0.047 446
#> mu_i[2]      18.791  0.654 17.533 18.769  20.107     0.03100          0.047 446
#> nu            3.532  1.223  2.078  3.225   6.674     0.06727          0.055 331
#>              R-hat
#> mu           0.999
#> sigma2       1.001
#> rho[1]       1.000
#> nu_p[1]      1.000
#> pooled_sigma 1.001
#> sigma_i[1]   0.999
#> sigma_i[2]   1.001
#> scale_i[1]   1.001
#> scale_i[2]   0.999
#> mu_i[1]      0.999
#> mu_i[2]      0.999
#> nu           1.000
#> Minimum effective sample size was low (160).
#> 
#>                                                                 
#>  Model               5             Parameter prior distributions
#>  Prior prob.     0.125                   delta ~ Cauchy(0, 0.71)
#>  log(marglik)   -75.75                     rho ~ Spike(0.5)     
#>  Post. prob.     0.017                                          
#>  Inclusion BF    0.120                                          
#> 
#> Parameter estimates:
#>                Mean    SD    lCI Median    uCI error(MCMC) error(MCMC)/SD ESS
#> mu           18.886 0.544 17.879 18.888 19.966     0.01891          0.035 828
#> sigma2        9.340 2.721  5.630  8.876 16.272     0.09697          0.036 787
#> delta[1]      0.701 0.369  0.018  0.701  1.454     0.01304          0.035 801
#> pooled_sigma  3.027 0.424  2.373  2.979  4.034     0.01475          0.035 826
#> sigma_i[1]    3.027 0.424  2.373  2.979  4.034     0.01475          0.035 826
#> sigma_i[2]    3.027 0.424  2.373  2.979  4.034     0.01475          0.035 826
#> mu_i[1]      17.848 0.753 16.327 17.867 19.340     0.02516          0.033 897
#> mu_i[2]      19.924 0.777 18.369 19.919 21.456     0.02629          0.034 874
#>              R-hat
#> mu           0.999
#> sigma2       1.001
#> delta[1]     0.999
#> pooled_sigma 1.001
#> sigma_i[1]   1.001
#> sigma_i[2]   1.001
#> mu_i[1]      0.999
#> mu_i[2]      0.999
#> 
#>                                                                 
#>  Model               6             Parameter prior distributions
#>  Prior prob.     0.125                   delta ~ Cauchy(0, 0.71)
#>  log(marglik)   -72.31                     rho ~ Spike(0.5)     
#>  Post. prob.     0.524                      nu ~ Exponential(1) 
#>  Inclusion BF    7.710                                          
#> 
#> Parameter estimates:
#>                Mean     SD    lCI Median     uCI error(MCMC) error(MCMC)/SD ESS
#> mu           19.595  0.400 18.706 19.621  20.278     0.02177          0.054 337
#> sigma2       25.158 41.099  4.893 13.833 113.827     2.52735          0.061 264
#> delta[1]      0.904  0.381  0.295  0.878   1.697     0.02235          0.059 290
#> nu_p[1]       0.771  0.831  0.046  0.526   2.934     0.04410          0.053 355
#> pooled_sigma  4.417  2.377  2.212  3.719  10.669     0.16038          0.067 220
#> sigma_i[1]    4.417  2.377  2.212  3.719  10.669     0.16038          0.067 220
#> sigma_i[2]    4.417  2.377  2.212  3.719  10.669     0.16038          0.067 220
#> scale_i[1]    1.710  0.372  1.095  1.666   2.508     0.01354          0.036 757
#> scale_i[2]    1.710  0.372  1.095  1.666   2.508     0.01354          0.036 757
#> mu_i[1]      17.904  0.557 16.753 17.920  18.930     0.02308          0.041 581
#> mu_i[2]      21.285  0.581 19.981 21.319  22.275     0.02737          0.047 451
#> nu            2.771  0.831  2.046  2.526   4.934     0.04410          0.053 355
#>              R-hat
#> mu           1.000
#> sigma2       1.000
#> delta[1]     1.006
#> nu_p[1]      1.004
#> pooled_sigma 1.004
#> sigma_i[1]   1.004
#> sigma_i[2]   1.004
#> scale_i[1]   0.999
#> scale_i[2]   0.999
#> mu_i[1]      1.000
#> mu_i[2]      0.999
#> nu           1.004
#> Minimum effective sample size was low (220).
#> 
#>                                                                 
#>  Model               7             Parameter prior distributions
#>  Prior prob.     0.125                   delta ~ Cauchy(0, 0.71)
#>  log(marglik)   -74.85                     rho ~ Beta(3, 3)     
#>  Post. prob.     0.041                                          
#>  Inclusion BF    0.301                                          
#> 
#> Parameter estimates:
#>                Mean    SD    lCI Median    uCI error(MCMC) error(MCMC)/SD ESS
#> mu           18.815 0.569 17.703 18.821 19.953     0.02118          0.037 721
#> sigma2        7.579 2.431  4.213  7.161 13.824     0.10366          0.043 550
#> delta[1]      0.716 0.372  0.028  0.704  1.490     0.01366          0.037 741
#> rho[1]        0.681 0.105  0.444  0.690  0.861     0.00440          0.042 568
#> pooled_sigma  3.018 0.436  2.308  2.962  4.035     0.01597          0.037 746
#> sigma_i[1]    2.370 0.510  1.636  2.289  3.611     0.02356          0.046 469
#> sigma_i[2]    3.515 0.609  2.570  3.418  4.909     0.01941          0.032 984
#> mu_i[1]      17.757 0.631 16.614 17.759 18.995     0.02061          0.033 937
#> mu_i[2]      19.874 0.906 18.164 19.883 21.606     0.03426          0.038 700
#>              R-hat
#> mu           0.999
#> sigma2       1.000
#> delta[1]     1.000
#> rho[1]       1.000
#> pooled_sigma 0.999
#> sigma_i[1]   1.000
#> sigma_i[2]   0.999
#> mu_i[1]      0.999
#> mu_i[2]      1.000
#> Minimum effective sample size was low (469).
#> 
#>                                                                 
#>  Model               8             Parameter prior distributions
#>  Prior prob.     0.125                   delta ~ Cauchy(0, 0.71)
#>  log(marglik)   -72.60                     rho ~ Beta(3, 3)     
#>  Post. prob.     0.393                      nu ~ Exponential(1) 
#>  Inclusion BF    4.530                                          
#> 
#> Parameter estimates:
#>                Mean     SD    lCI Median     uCI error(MCMC) error(MCMC)/SD ESS
#> mu           19.580  0.399 18.769 19.602  20.353     0.01869          0.047 455
#> sigma2       21.138 31.637  4.555 12.028 107.607     2.45693          0.078 166
#> delta[1]      0.910  0.375  0.288  0.875   1.751     0.02166          0.058 300
#> rho[1]        0.555  0.134  0.292  0.558   0.798     0.00522          0.039 657
#> nu_p[1]       0.826  0.806  0.038  0.579   3.004     0.03735          0.046 466
#> pooled_sigma  4.271  2.157  2.174  3.633  10.444     0.15503          0.072 194
#> sigma_i[1]    3.996  2.170  1.905  3.364  10.365     0.14920          0.069 212
#> sigma_i[2]    4.451  2.304  2.131  3.831  10.922     0.16267          0.071 201
#> scale_i[1]    1.628  0.415  0.989  1.561   2.611     0.01437          0.035 832
#> scale_i[2]    1.849  0.523  0.999  1.793   2.955     0.01879          0.036 773
#> mu_i[1]      17.904  0.532 16.867 17.919  18.966     0.01942          0.036 751
#> mu_i[2]      21.256  0.593 20.037 21.294  22.345     0.02601          0.044 520
#> nu            2.826  0.806  2.038  2.579   5.004     0.03735          0.046 466
#>              R-hat
#> mu           1.002
#> sigma2       1.009
#> delta[1]     0.999
#> rho[1]       1.000
#> nu_p[1]      0.999
#> pooled_sigma 1.005
#> sigma_i[1]   1.006
#> sigma_i[2]   1.005
#> scale_i[1]   0.999
#> scale_i[2]   1.000
#> mu_i[1]      1.001
#> mu_i[2]      1.000
#> nu           0.999
#> Minimum effective sample size was low (166).