Summarize fitted 'RoBTT' object

summary.RoBTT creates summary tables for a RoBTT object.

# S3 method for RoBTT
summary(
  object,
  type = "ensemble",
  conditional = FALSE,
  group_estimates = FALSE,
  probs = c(0.025, 0.975),
  logBF = FALSE,
  BF01 = FALSE,
  short_name = FALSE,
  remove_spike_0 = FALSE,
  ...
)

Arguments

object

a fitted 'RoBTT' object

type

whether to show the overall 'RoBTT' results ("ensemble"), an overview of the individual models ("models"), an overview of the individual models MCMC diagnostics ("diagnostics"), or a detailed summary of the individual models ("individual"). Can be abbreviated to first letters.

conditional

show the conditional estimates (assuming that the alternative is true). Defaults to FALSE. Only available for type == "conditional".

group_estimates

show the model-averaged mean and standard deviation estimates for each group.

probs

quantiles of the posterior samples to be displayed. Defaults to c(.025, .975)

logBF

show log of Bayes factors. Defaults to FALSE.

BF01

show Bayes factors in support of the null hypotheses. Defaults to FALSE.

short_name

whether priors names should be shortened to the first (couple) of letters. Defaults to FALSE.

remove_spike_0

whether spike prior distributions with location at zero should be omitted from the summary. Defaults to FALSE.

...

additional arguments

Value

summary.RoBTT returns a list of tables of class 'BayesTools_table'.

See also

Examples

# using the example data from Darwin
data("fertilization", package = "RoBTT")
fit <- RoBTT(
  x1       = fertilization$Self,
  x2       = fertilization$Crossed,
  prior_delta = prior("cauchy", list(0, 1/sqrt(2))),
  prior_rho   = prior("beta",   list(3, 3)),
  likelihood  = "normal",
  seed        = 1, 
  chains      = 1,
  warmup      = 1000,
  iter        = 2000,
  control     = set_control(adapt_delta = 0.95)
)
#> Chain 1: 
#> Chain 1: Gradient evaluation took 1.1e-05 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 0.11 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: 
#> Chain 1:  Elapsed Time: 0.020363 seconds (Warm-up)
#> Chain 1:                0.018033 seconds (Sampling)
#> Chain 1:                0.038396 seconds (Total)
#> Chain 1: 
#> 0 of 1000 iterations ended with a divergence.
#> Chain 1: 
#> Chain 1: Gradient evaluation took 1.3e-05 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 0.13 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: 
#> Chain 1:  Elapsed Time: 0.035097 seconds (Warm-up)
#> Chain 1:                0.038634 seconds (Sampling)
#> Chain 1:                0.073731 seconds (Total)
#> Chain 1: 
#> 0 of 1000 iterations ended with a divergence.
#> Chain 1: 
#> Chain 1: Gradient evaluation took 1.2e-05 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 0.12 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: 
#> Chain 1:  Elapsed Time: 0.026382 seconds (Warm-up)
#> Chain 1:                0.024467 seconds (Sampling)
#> Chain 1:                0.050849 seconds (Total)
#> Chain 1: 
#> 0 of 1000 iterations ended with a divergence.
#> Chain 1: 
#> Chain 1: Gradient evaluation took 1.5e-05 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 0.15 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: 
#> Chain 1:  Elapsed Time: 0.04049 seconds (Warm-up)
#> Chain 1:                0.039786 seconds (Sampling)
#> Chain 1:                0.080276 seconds (Total)
#> Chain 1: 
#> 0 of 1000 iterations ended with a divergence.
#> Warning: Model (2): Minimum effective sample size was low (325).
#> Warning: Model (4): Minimum effective sample size was low (469).

# summary can provide many details about the model
summary(fit)
#> Call:
#> RoBTT(x1 = fertilization$Self, x2 = fertilization$Crossed, prior_delta = prior("cauchy", 
#>     list(0, 1/sqrt(2))), prior_rho = prior("beta", list(3, 3)), 
#>     likelihood = "normal", chains = 1, iter = 2000, warmup = 1000, 
#>     control = set_control(adapt_delta = 0.95), seed = 1)
#> 
#> Robust Bayesian t-test
#> Components summary:
#>               Models Prior prob. Post. prob. Inclusion BF
#> Effect           2/4       0.500       0.750        3.002
#> Heterogeneity    2/4       0.500       0.709        2.431
#> 
#> Model-averaged estimates:
#>        Mean Median 0.025 0.975
#> delta 0.540  0.546 0.000 1.425
#> rho   0.632  0.637 0.469 0.849
#> Model (2): Minimum effective sample size was low (325).
#> Model (4): Minimum effective sample size was low (469).

# estimates from the conditional models can be obtained with
summary(fit, conditional = TRUE)
#> Call:
#> RoBTT(x1 = fertilization$Self, x2 = fertilization$Crossed, prior_delta = prior("cauchy", 
#>     list(0, 1/sqrt(2))), prior_rho = prior("beta", list(3, 3)), 
#>     likelihood = "normal", chains = 1, iter = 2000, warmup = 1000, 
#>     control = set_control(adapt_delta = 0.95), seed = 1)
#> 
#> Robust Bayesian t-test
#> Components summary:
#>               Models Prior prob. Post. prob. Inclusion BF
#> Effect           2/4       0.500       0.750        3.002
#> Heterogeneity    2/4       0.500       0.709        2.431
#> 
#> Model-averaged estimates:
#>        Mean Median 0.025 0.975
#> delta 0.540  0.546 0.000 1.425
#> rho   0.632  0.637 0.469 0.849
#> Model (2): Minimum effective sample size was low (325).
#> Model (4): Minimum effective sample size was low (469).
#> 
#> Conditional estimates:
#>        Mean Median 0.025 0.975
#> delta 0.712  0.701 0.028 1.467
#> rho   0.688  0.698 0.440 0.863
#> Model (2): Minimum effective sample size was low (325).
#> Model (4): Minimum effective sample size was low (469).

# overview of the models and their prior and posterior probability, marginal likelihood,
# and inclusion Bayes factor can be obtained with
summary(fit, type = "models")
#> Call:
#> RoBTT(x1 = fertilization$Self, x2 = fertilization$Crossed, prior_delta = prior("cauchy", 
#>     list(0, 1/sqrt(2))), prior_rho = prior("beta", list(3, 3)), 
#>     likelihood = "normal", chains = 1, iter = 2000, warmup = 1000, 
#>     control = set_control(adapt_delta = 0.95), seed = 1)
#> 
#> Robust Bayesian t-test
#> Models overview:
#>  Model Distribution   Prior delta    Prior rho Prior prob. log(marglik)
#>      1       normal        Spike(0) Spike(0.5)       0.250       -76.82
#>      2       normal        Spike(0) Beta(3, 3)       0.250       -75.96
#>      3       normal Cauchy(0, 0.71) Spike(0.5)       0.250       -75.75
#>      4       normal Cauchy(0, 0.71) Beta(3, 3)       0.250       -74.85
#>  Post. prob. Inclusion BF
#>        0.074        0.240
#>        0.176        0.640
#>        0.217        0.833
#>        0.533        3.420
#> Model (2): Minimum effective sample size was low (325).
#> Model (4): Minimum effective sample size was low (469).

# diagnostics overview, containing the maximum R-hat, minimum ESS, maximum MCMC error, and
# maximum MCMC error / sd across parameters for each individual model can be obtained with
summary(fit, type = "diagnostics")
#> Call:
#> RoBTT(x1 = fertilization$Self, x2 = fertilization$Crossed, prior_delta = prior("cauchy", 
#>     list(0, 1/sqrt(2))), prior_rho = prior("beta", list(3, 3)), 
#>     likelihood = "normal", chains = 1, iter = 2000, warmup = 1000, 
#>     control = set_control(adapt_delta = 0.95), seed = 1)
#> 
#> Robust Bayesian t-test
#> Diagnostics overview:
#>  Model Distribution   Prior delta    Prior rho max[error(MCMC)]
#>      1       normal        Spike(0) Spike(0.5)          0.13670
#>      2       normal        Spike(0) Beta(3, 3)          0.14611
#>      3       normal Cauchy(0, 0.71) Spike(0.5)          0.09697
#>      4       normal Cauchy(0, 0.71) Beta(3, 3)          0.10366
#>  max[error(MCMC)/SD] min(ESS) max(R-hat)
#>                0.043      549      1.000
#>                0.055      325      1.007
#>                0.036      787      1.001
#>                0.046      469      1.000
#> Model (2): Minimum effective sample size was low (325).
#> Model (4): Minimum effective sample size was low (469).

# summary of individual models and their parameters can be further obtained by
summary(fit, type = "individual")
#> Call:
#> RoBTT(x1 = fertilization$Self, x2 = fertilization$Crossed, prior_delta = prior("cauchy", 
#>     list(0, 1/sqrt(2))), prior_rho = prior("beta", list(3, 3)), 
#>     likelihood = "normal", chains = 1, iter = 2000, warmup = 1000, 
#>     control = set_control(adapt_delta = 0.95), seed = 1)
#> 
#> Robust Bayesian t-test                                                                
#>  Model               1             Parameter prior distributions
#>  Prior prob.     0.250                        delta ~ Spike(0)  
#>  log(marglik)   -76.82                          rho ~ Spike(0.5)
#>  Post. prob.     0.074                                          
#>  Inclusion BF    0.240                                          
#> 
#> Parameter estimates:
#>                Mean    SD    lCI Median    uCI error(MCMC) error(MCMC)/SD ESS
#> mu           18.890 0.590 17.746 18.876 20.058     0.02222          0.038 706
#> sigma2       10.981 3.203  6.380 10.348 18.176     0.13670          0.043 549
#> pooled_sigma  3.281 0.463  2.526  3.217  4.263     0.01960          0.042 558
#> sigma_i[1]    3.281 0.463  2.526  3.217  4.263     0.01960          0.042 558
#> sigma_i[2]    3.281 0.463  2.526  3.217  4.263     0.01960          0.042 558
#> mu_i[1]      18.890 0.590 17.746 18.876 20.058     0.02222          0.038 706
#> mu_i[2]      18.890 0.590 17.746 18.876 20.058     0.02222          0.038 706
#>              R-hat
#> mu           1.000
#> sigma2       1.000
#> pooled_sigma 1.000
#> sigma_i[1]   1.000
#> sigma_i[2]   1.000
#> mu_i[1]      1.000
#> mu_i[2]      1.000
#> 
#>                                                                 
#>  Model               2             Parameter prior distributions
#>  Prior prob.     0.250                        delta ~ Spike(0)  
#>  log(marglik)   -75.96                          rho ~ Beta(3, 3)
#>  Post. prob.     0.176                                          
#>  Inclusion BF    0.640                                          
#> 
#> Parameter estimates:
#>                Mean    SD    lCI Median    uCI error(MCMC) error(MCMC)/SD ESS
#> mu           18.348 0.606 17.248 18.331 19.571     0.02630          0.043 532
#> sigma2        8.556 2.835  4.739  8.099 15.212     0.14611          0.052 376
#> rho[1]        0.698 0.109  0.436  0.711  0.872     0.00598          0.055 335
#> pooled_sigma  3.273 0.483  2.526  3.195  4.392     0.01858          0.038 676
#> sigma_i[1]    2.489 0.552  1.689  2.404  3.754     0.03060          0.055 325
#> sigma_i[2]    3.860 0.698  2.784  3.752  5.492     0.02363          0.034 873
#> mu_i[1]      18.348 0.606 17.248 18.331 19.571     0.02630          0.043 532
#> mu_i[2]      18.348 0.606 17.248 18.331 19.571     0.02630          0.043 532
#>              R-hat
#> mu           1.001
#> sigma2       1.007
#> rho[1]       1.003
#> pooled_sigma 0.999
#> sigma_i[1]   1.006
#> sigma_i[2]   0.999
#> mu_i[1]      1.001
#> mu_i[2]      1.001
#> Minimum effective sample size was low (325).
#> 
#>                                                                 
#>  Model               3             Parameter prior distributions
#>  Prior prob.     0.250                   delta ~ Cauchy(0, 0.71)
#>  log(marglik)   -75.75                     rho ~ Spike(0.5)     
#>  Post. prob.     0.217                                          
#>  Inclusion BF    0.833                                          
#> 
#> Parameter estimates:
#>                Mean    SD    lCI Median    uCI error(MCMC) error(MCMC)/SD ESS
#> mu           18.886 0.544 17.879 18.888 19.966     0.01891          0.035 828
#> sigma2        9.340 2.721  5.630  8.876 16.272     0.09697          0.036 787
#> delta[1]      0.701 0.369  0.018  0.701  1.454     0.01304          0.035 801
#> pooled_sigma  3.027 0.424  2.373  2.979  4.034     0.01475          0.035 826
#> sigma_i[1]    3.027 0.424  2.373  2.979  4.034     0.01475          0.035 826
#> sigma_i[2]    3.027 0.424  2.373  2.979  4.034     0.01475          0.035 826
#> mu_i[1]      17.848 0.753 16.327 17.867 19.340     0.02516          0.033 897
#> mu_i[2]      19.924 0.777 18.369 19.919 21.456     0.02629          0.034 874
#>              R-hat
#> mu           0.999
#> sigma2       1.001
#> delta[1]     0.999
#> pooled_sigma 1.001
#> sigma_i[1]   1.001
#> sigma_i[2]   1.001
#> mu_i[1]      0.999
#> mu_i[2]      0.999
#> 
#>                                                                 
#>  Model               4             Parameter prior distributions
#>  Prior prob.     0.250                   delta ~ Cauchy(0, 0.71)
#>  log(marglik)   -74.85                     rho ~ Beta(3, 3)     
#>  Post. prob.     0.533                                          
#>  Inclusion BF    3.420                                          
#> 
#> Parameter estimates:
#>                Mean    SD    lCI Median    uCI error(MCMC) error(MCMC)/SD ESS
#> mu           18.815 0.569 17.703 18.821 19.953     0.02118          0.037 721
#> sigma2        7.579 2.431  4.213  7.161 13.824     0.10366          0.043 550
#> delta[1]      0.716 0.372  0.028  0.704  1.490     0.01366          0.037 741
#> rho[1]        0.681 0.105  0.444  0.690  0.861     0.00440          0.042 568
#> pooled_sigma  3.018 0.436  2.308  2.962  4.035     0.01597          0.037 746
#> sigma_i[1]    2.370 0.510  1.636  2.289  3.611     0.02356          0.046 469
#> sigma_i[2]    3.515 0.609  2.570  3.418  4.909     0.01941          0.032 984
#> mu_i[1]      17.757 0.631 16.614 17.759 18.995     0.02061          0.033 937
#> mu_i[2]      19.874 0.906 18.164 19.883 21.606     0.03426          0.038 700
#>              R-hat
#> mu           0.999
#> sigma2       1.000
#> delta[1]     1.000
#> rho[1]       1.000
#> pooled_sigma 0.999
#> sigma_i[1]   1.000
#> sigma_i[2]   0.999
#> mu_i[1]      0.999
#> mu_i[2]      1.000
#> Minimum effective sample size was low (469).